Improve perk search
